backup_before_cleanup/mobile-agentx/requirements.txt
# Mobile AgentX Requirements
# Core dependencies for the mobile automation platform

# AgentSphere SDK (your existing framework)
google-adk

# Data modeling and validation
pydantic>=2.0.0

# Date/time handling
python-dateutil>=2.8.0

# Optional: For real API integrations (when you move from mock mode)
requests>=2.31.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.88.0

# Optional: For enhanced mobile features
geopy>=2.3.0  # For location services
pillow>=10.0.0  # For image handling